Cleanup IPAM usage in the util package
IPAM was planned to be used in SMF. Specially it was useful
when multiple SMF instances need to acquire IP address and
this module helps to avoid the duplicate assignment.

But long term having IP address allocation at UPF is best strategy
as it helps in distributing ip address assigmne to multiple endpoints
and this reduces the burden at SMF.

This is first step in simplifying the implementation.

Signed-off-by: Ajay Lotan Thakur <thakur.ajay@gmail.com>